requirements.txt
discord.py
asyncio
aiohttp
pprint
configparser
pillow